refactor: rename SiteHeader to Header
