+++ src/index.jsx
@@ -1,6 +1,6 @@
-import SiteHeader from './SiteHeader';
+import Header from './Header';
 import messages from './i18n/index';
 
 export { messages };
 
-export default SiteHeader;
+export default Header;
